Backend (RAG Model)/requirements.txt
cohere
pinecone-client
transformers
flask
python-dotenv
PyPDF2
